Pass in the actor on the command line.
GITHUB_ACTOR is the GitHub account which triggered the CI
Runner, not the GitHub App we created for this. For example
if I submitted a PR the GITHUB_ACTOR would be my GitHub account,
not the acount created for the bot.

Pass in the actor explicitly on the commandline instead.

Signed-off-by: Denton Gentry <dgentry@tailscale.com>