Move configuration out of YAML and into secrets.
Preparing to open the source: forking the repo should not result
in someone else running a GitHub Action every few minutes watching
Tailscale's OSS repo and filing bugs. Even though the issues they
file would be in their own repo's issue tracker, it would show up
as mentions on our PRs and that would be annoying.

Move the configuration out of the YAML file and into secrets.
Forking the repo won't fork the secrets, a forked repo will
run the action periodically but not do anything.

Since the Actions runner minutes will be free as a public repo,
run the bot a little more often.

Signed-off-by: Denton Gentry <dgentry@tailscale.com>